news_agent/requirements.txt
hive-agent @ git+https://github.com/hivenetwork-ai/hive-agent-py@main
aiosqlite==0.20.0
python-dotenv==1.0.1
requests==2.31.0
newsapi-python==0.2.7
